+++ stapi-fastapi/src/stapi_fastapi/__init__.py
@@ -1,6 +1,8 @@
+from .pagination import Page
 from .routers import ProductRouter, RootRouter
 
 __all__ = [
+    "Page",
     "ProductRouter",
     "RootRouter",
 ]